+++ Nefarius.DSharpPlus.Extensions.Generators/DiscordClientEventsInterfacesGenerator.cs
@@ -1,42 +1,22 @@
 ﻿using System.Linq;
-using System.Net;
 using System.Text;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
+using Nefarius.DSharpPlus.Extensions.Generators.Util;
 
-namespace Nefarius.DSharpPlus.Extensions.Hosting.Generators
+namespace Nefarius.DSharpPlus.Extensions.Generators
 {
     [Generator]
     public class DiscordClientEventsInterfacesGenerator : ISourceGenerator
     {
-        private const string DSharpPlusSourceUri =
-            "https://raw.githubusercontent.com/DSharpPlus/DSharpPlus/master/DSharpPlus/Clients/DiscordClient.Events.cs";
-
         public void Initialize(GeneratorInitializationContext context)
         {
         }
 
         public void Execute(GeneratorExecutionContext context)
         {
-            using var client = new WebClient();
-
-            // Required or result is HTTP-403
-            client.Headers["User-Agent"] =
-                "Mozilla/4.0 (Compatible; Windows NT 5.1; MSIE 6.0) " +
-                "(compatible; MSIE 6.0; Windows NT 5.1; " +
-                ".NET CLR 1.1.4322; .NET CLR 2.0.50727)";
-
-            var response = client.DownloadString(DSharpPlusSourceUri);
-
-            var syntaxTree = CSharpSyntaxTree.ParseText(response);
-
-            var root = syntaxTree.GetCompilationUnitRoot();
-
-            var namespaceSyntax = root.Members.OfType<NamespaceDeclarationSyntax>().First();
-
-            var discordClientClassSyntax = namespaceSyntax.Members.OfType<ClassDeclarationSyntax>().First();
+            var discordClientClassSyntax = DSharpPlusClientParser.Instance.DiscordClient;
 
             var eventsSyntax = discordClientClassSyntax.Members.OfType<EventDeclarationSyntax>();
